Only delete .bvec/.bval files if all bvals are (near) zero.
Print out a warning if they are saved in destination folder if it is not `dwi`

+++ heudiconv/convert.py
@@ -679,9 +679,17 @@ def save_converted_files(res, item_dicoms, bids_options, outtype, prefix, outnam
             safe_movefile(res.outputs.bvecs, outname_bvecs, overwrite)
             safe_movefile(res.outputs.bvals, outname_bvals, overwrite)
         else:
-            os.remove(res.outputs.bvecs)
-            os.remove(res.outputs.bvals)
-            lgr.debug("%s and %s were removed since not dwi", res.outputs.bvecs, res.outputs.bvals)
+            if bvals_are_zero(res.outputs.bvals):
+                os.remove(res.outputs.bvecs)
+                os.remove(res.outputs.bvals)
+                lgr.debug("%s and %s were removed since not dwi", res.outputs.bvecs, res.outputs.bvals)
+            else:
+                lgr.info("Diffusion-weighted image saved in non dwi folder (%s)", prefix_dirname)
+                lgr.info(".bvec and .bval files will be generated. This is NOT BIDS compliant")
+                outname_bvecs, outname_bvals = prefix + '.bvec', prefix + '.bval'
+                safe_movefile(res.outputs.bvecs, outname_bvecs, overwrite)
+                safe_movefile(res.outputs.bvals, outname_bvals, overwrite)
+
 
     if isinstance(res_files, list):
         res_files = sorted(res_files)
@@ -811,3 +819,23 @@ def  add_taskname_to_infofile(infofiles):
 
         # write to outfile
         save_json(infofile, meta_info)
+
+
+def bvals_are_zero(bval_file, threshold=5):
+    """Checks if all entries in a bvals file are zero (or below the threshold).
+    Returns True if that is the case, otherwise returns False
+
+    Parameters
+    ----------
+    bval_file : file with the bvals
+    threshold : b-value below which they are considered to be zero. (Default: 5)
+
+    Returns
+    -------
+    True if all are zero; False otherwise.
+    """
+
+    with open(bval_file) as f:
+        bvals = f.read().split()
+
+    return all(float(b) <= threshold for b in bvals)
